updating: split installer-related and github-related APIs, fix compilation error with latest VS
