+++ installer/PowerToysBootstrapper/bootstrapper/pch.h
@@ -20,4 +20,8 @@
 #include <spdlog/sinks/null_sink.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 
+#include <winrt/base.h>
+#include <winrt/Windows.Foundation.h>
+#include <winrt/Windows.Foundation.Collections.h>
+
 #include <cxxopts.hpp>
