+++ src/action_runner/action_runner.cpp
@@ -4,10 +4,12 @@
 #include <Windows.h>
 #include <shellapi.h>
 
+#include <filesystem>
 #include <string_view>
 
 #include <common/common.h>
 #include <common/updating/updating.h>
+#include <common/updating/installer.h>
 #include <common/updating/http_client.h>
 #include <common/updating/dotnet_installation.h>
