+++ src/common/updating/updating.h
@@ -11,14 +11,6 @@
 
 namespace updating
 {
-    std::wstring get_msi_package_path();
-    bool uninstall_msi_version(const std::wstring& package_path, const notifications::strings&);
-    bool offer_msi_uninstallation(const notifications::strings&);
-    std::optional<std::wstring> get_msi_package_installed_path();
-    std::optional<VersionHelper> get_installed_powertoys_version();
-
-    std::future<bool> uninstall_previous_msix_version_async();
-
     struct new_version_download_info
     {
         winrt::Windows::Foundation::Uri release_page_uri;
