+++ src/common/updating/updating.vcxproj.filters
@@ -30,6 +30,9 @@
     <ClInclude Include="notifications.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="installer.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp">
@@ -45,6 +48,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="notifications.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="installer.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
